task: Remove Exploration mission class
Changes:
- Move ExplorationMeta and ExplorationMetaTest to predefined mission package
- Remove Exploration mission class
- Move the SITE_EXPLORATION_EVENT & REQUIRED_SPECIMEN_CONTAINER constants to ExplorationMeta
- Updated users of constants

ref #975
