+++ mars-sim-ui/src/main/java/com/mars_sim/ui/swing/entitywindow/mission/TabPanelNavigation.java
@@ -25,8 +25,6 @@ import com.mars_sim.core.EntityEvent;
 import com.mars_sim.core.EntityEventType;
 import com.mars_sim.core.EntityListener;
 import com.mars_sim.core.map.MapDataFactory;
-import com.mars_sim.core.person.ai.mission.Exploration;
-import com.mars_sim.core.person.ai.mission.Mining;
 import com.mars_sim.core.person.ai.mission.NavPoint;
 import com.mars_sim.core.person.ai.mission.VehicleMission;
 import com.mars_sim.core.tool.Msg;
@@ -35,7 +33,6 @@ import com.mars_sim.ui.swing.ImageLoader;
 import com.mars_sim.ui.swing.UIContext;
 import com.mars_sim.ui.swing.entitywindow.EntityTabPanel;
 import com.mars_sim.ui.swing.tool.map.MapPanel;
-import com.mars_sim.ui.swing.tool.map.MineralMapLayer;
 import com.mars_sim.ui.swing.tool.map.MissionMapLayer;
 import com.mars_sim.ui.swing.tool.map.UnitMapLayer;
 import com.mars_sim.ui.swing.tool.map.VehicleTrailMapLayer;
@@ -98,12 +95,6 @@ class TabPanelNavigation extends EntityTabPanel<VehicleMission>
 		mapPanel = new MapPanel(getContext());
 		mapPanel.setBackground(new Color(0, 0, 0, 128));
 		mapPanel.setOpaque(false);
-		
-		// Mineral layer only if mining or exploration mission
-		if (mission instanceof Mining || mission instanceof Exploration) {
-		    var mineralLayer = new MineralMapLayer(mapPanel);
-		    mapPanel.addMapLayer(mineralLayer);
-		}					
 
 		// Always add unit layer
 		mapPanel.addMapLayer(new UnitMapLayer(mapPanel));
