+++ mars-sim-ui/src/main/java/com/mars_sim/ui/swing/tool/missionwizard/MissionDataBean.java
@@ -17,6 +17,7 @@ import com.mars_sim.core.goods.Good;
 import com.mars_sim.core.map.location.Coordinates;
 import com.mars_sim.core.mission.MetaMission;
 import com.mars_sim.core.mission.MetaMissionRegistry;
+import com.mars_sim.core.mission.predefined.ExplorationMeta;
 import com.mars_sim.core.mission.predefined.LandmarkMetaMission;
 import com.mars_sim.core.mission.predefined.TestDriveMetaMission;
 import com.mars_sim.core.person.Person;
@@ -29,7 +30,6 @@ import com.mars_sim.core.person.ai.mission.meta.CollectRegolithMeta;
 import com.mars_sim.core.person.ai.mission.meta.ConstructionMissionMeta;
 import com.mars_sim.core.person.ai.mission.meta.DeliveryMeta;
 import com.mars_sim.core.person.ai.mission.meta.EmergencySupplyMeta;
-import com.mars_sim.core.person.ai.mission.meta.ExplorationMeta;
 import com.mars_sim.core.person.ai.mission.meta.MeteorologyFieldStudyMeta;
 import com.mars_sim.core.person.ai.mission.meta.MiningMeta;
 import com.mars_sim.core.person.ai.mission.meta.RescueSalvageVehicleMeta;
